03/01/25 OgreWin3D_Stage :- Room Builder Brush Selection

+++ Room_Builder/CL64_Array.h
@@ -10,7 +10,7 @@ struct tag_Array
 
 #define Array_GetSize(a) ((a)->ItemsAllocated)
 #define Array_GetItemSize(a) ((a)->ItemSize)
-#define Array_ItemPtr(a,i) ((void *)(((long)((a)->Items))+((i)*(a)->ItemSize)))
+#define Array_ItemPtr(a,i) ((void *)(((uintptr_t)((a)->Items))+((i)*(a)->ItemSize)))
 #define Array_PutAt(a,i,d,s) (memcpy (Array_ItemPtr((a),(i)),(d),(s)))
 #define Array_InsertAt(a,i,d,s) \
 		(memmove(Array_ItemPtr((a),(i)+1),Array_ItemPtr((a),(i)),((a)->ItemsAllocated-(i)-1)*(a)->ItemSize), \
